refactor(cockpit): extract risk module
